rename ConversionTypes.h -> BankSelectStyle.h
